fix(completion): Don't print duplicate primitive modules
