+++ src/lib.rs
@@ -49,6 +49,7 @@ use std::result::Result as StdResult;
 use std::string::String as StdString;
 use std::env;
 
+use base::filename_to_module;
 use base::ast::{self, SpannedExpr};
 use base::error::{Errors, InFile};
 use base::metadata::Metadata;
@@ -386,7 +387,7 @@ impl Compiler {
                 &owned_import
             }
         };
-        let module_name = Symbol::from(filename_to_module(filename));
+        let module_name = Symbol::from(format!("@{}", filename_to_module(filename)));
         let mut macros = MacroExpander::new(vm);
         FutureValue::from(
             import
@@ -549,18 +550,6 @@ let { error } = import! std.prim
 
 in ()
 "#;
-
-pub fn filename_to_module(filename: &str) -> StdString {
-    use std::path::Path;
-    let path = Path::new(filename);
-    let name = path.extension().map_or(filename, |ext| {
-        ext.to_str()
-            .map(|ext| &filename[..filename.len() - ext.len() - 1])
-            .unwrap_or(filename)
-    });
-
-    format!("@{}", name.replace(|c: char| c == '/' || c == '\\', "."))
-}
 
 #[derive(Default)]
 pub struct VmBuilder {
